Hadled duplicate and add only new items.
