+++ src/main/java/org/jtgm/core/service/impl/DefaultExcelExtractor.java
@@ -1,14 +1,15 @@
-package org.jtgm.core.service;
+package org.jtgm.core.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.poi.ss.usermodel.*;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.jtgm.core.service.ExcelExtractor;
 import org.jtgm.core.util.ExcelUtil;
 import org.jtgm.core.exception.GenericErrorException;
 import org.springframework.web.multipart.MultipartFile;
 
 @RequiredArgsConstructor
-public class DefaultExcelExtractor implements ExcelExtractor{
+public class DefaultExcelExtractor implements ExcelExtractor {
     private final ExcelUtil excelUtil;
 
     @Override
@@ -22,7 +23,7 @@ public class DefaultExcelExtractor implements ExcelExtractor{
             excelUtil.execute(sheet, mgroupName);
         }catch (Exception e) {
             e.printStackTrace();
-            throw new GenericErrorException("Error occur", e);
+            throw new GenericErrorException("Unable to process file", e);
         }
     }
