--- src/main/java/org/jtgm/core/util/DefaultValidationUtil.java
@@ -1,11 +1,0 @@
-package org.jtgm.core.util;
-
-import org.apache.poi.ss.usermodel.Sheet;
-
-public class DefaultValidationUtil implements ValidationUtil{
-
-    @Override
-    public void validate(Sheet sheet) {
-
-    }
-}
