+++ src/main/java/org/jtgm/presentation/ExcelController.java
@@ -20,9 +20,9 @@ public class ExcelController {
     @PostMapping(path = "/mgroup",  produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> saveMgroup(@RequestPart(value="file") MultipartFile file) {
         try{
-            log.info("[START] Hitting a test endpoint");
+            log.info("[START] Initiating process of MGroup files to staging excel.");
             excelExtractor.extract(file);
-            log.info("[END] Hitting a test endpoint");
+            log.info("[END] Done processing the MGroup files to staging excel.");
             return new ResponseEntity<>("Extraction finished", HttpStatus.OK);
         }catch (Exception ex){
             return new ResponseEntity<>("Failed to extract the excel file", HttpStatus.INTERNAL_SERVER_ERROR);
